+++ t/029_union_pins.t
@@ -1,19 +1,19 @@
 use v5.40;
-use Affix qw[:all];
-use Test2::V0;
+use Affix               qw[:all];
+use Test2::Tools::Affix qw[:all];
 typedef MyUnion => Union [ i => Int, f => Float ];
 
 # Allocate some memory for the union
-my $mem = malloc(sizeof(MyUnion()));
-my $u = cast($mem, MyUnion());
+my $mem = malloc( sizeof( MyUnion() ) );
+my $u   = cast( $mem, MyUnion() );
 
 # $u should be a hashref with Pins
-ok ref($u) eq 'HASH', 'Union is a HASH';
+ok ref($u) eq 'HASH',  'Union is a HASH';
 ok is_pin( \$u->{i} ), 'member i is a Pin';
 ok is_pin( \$u->{f} ), 'member f is a Pin';
 
 # Write to one, check the other
-$u->{i} = 0x3f800000; # 1.0 in float
+$u->{i} = 0x3f800000;    # 1.0 in float
 is $u->{f}, float(1.0), 'Writing to i affected f';
 
 # Write to f, check i